+++ lambda/service/store/application.go
@@ -1,4 +1,4 @@
-package repository
+package store
 
 import "time"
